gynecology_chatbot/backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
distro==1.9.0
Django==5.2.3
django-allauth==65.9.0
django-cors-headers==4.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dnspython==2.4.2
drf-nested-routers==0.93.4
exceptiongroup==1.3.0
factory-boy==3.3.0
Faker==37.3.0
firebase-admin==6.4.0
frozenlist==1.6.2
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc1
google-api-python-client==2.170.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-core==2.4.3
google-cloud-firestore==2.14.0
google-cloud-storage==2.10.0
google-crc32c==1.7.1
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jiter==0.10.0
mongoengine==0.27.0
msgpack==1.1.0
multidict==6.4.4
oauthlib==3.2.2
openai==1.84.0
packaging==25.0
pluggy==1.6.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==4.25.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.9.0
pymongo==4.6.0
pyparsing==3.2.3
pytest==7.4.4
pytest-asyncio==0.23.5
pytest-django==4.8.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tomli==2.2.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
yarl==1.20.0


# Django==5.2.2
# djangorestframework==3.15.2
# django-cors-headers==4.1.0
# djangorestframework-simplejwt==5.3.2
# drf-nested-routers==0.93.4
# python-dotenv==1.0.0
# openai==1.84.0
# google-generativeai>=0.8.0
# google-cloud-firestore==2.14.0
# google-cloud-storage==2.10.0
# firebase-admin==6.4.0